requirements.txt
appdirs==1.4.0
Django==1.8.6
django-crispy-forms==1.6.1
django-taggit==0.17.1
django-widget-tweaks==1.4.1
packaging==16.8
Pillow==2.9.0
pyparsing==2.1.10
pytz==2016.10
redis==2.10.3
six==1.10.0
sorl-thumbnail==12.3
